+++ ForgeMenuChoice/ForgeMenuChoice/HarmonyPatches/ForgeMenuPatches.cs
@@ -11,6 +11,7 @@ using StardewModdingAPI.Utilities;
 
 using StardewValley.Enchantments;
 using StardewValley.Menus;
+using StardewValley.Tools;
 
 namespace ForgeMenuChoice.HarmonyPatches;
 
@@ -89,15 +90,20 @@ internal static class ForgeMenuPatches
     [HarmonyPrefix]
     [HarmonyPriority(Priority.High)]
     [HarmonyPatch(nameof(ForgeMenu.IsValidCraft))]
-    internal static bool PrefixIsValidCraft(Item __0, Item __1, ref bool __result)
+    internal static bool PrefixIsValidCraft(Item __0, Item __1, ref bool __result, ForgeMenu __instance)
     {
         try
         {
             // 74 - prismatic shard.
             if (__0 is Tool tool && __1.QualifiedItemId == "(O)74")
             {
+                if (Menu is not null && !Menu.IsInnate && ReferenceEquals(Menu.Tool, tool))
+                {
+                    __result = true;
+                    return false;
+                }
                 PossibleEnchantments.Clear();
-                HashSet<Type> enchants = tool.enchantments.Select((a) => a.GetType()).ToHashSet();
+                HashSet<Type> enchants = tool.enchantments.Select(static a => a.GetType()).ToHashSet();
                 foreach (BaseEnchantment enchantment in BaseEnchantment.GetAvailableEnchantments())
                 {
                     if (enchantment.CanApplyTo(tool) && !enchants.Contains(enchantment.GetType()))
@@ -107,7 +113,27 @@ internal static class ForgeMenuPatches
                 }
                 if (PossibleEnchantments.Count > 0)
                 {
-                    Menu ??= new(options: PossibleEnchantments);
+                    Menu = new(options: PossibleEnchantments, tool, false);
+                    __result = true;
+                    return false;
+                }
+            }
+            else if (ModEntry.Config.OverrideInnateEnchantments && ReferenceEquals(__1, __instance.rightIngredientSpot.item)
+                && __0 is MeleeWeapon weapon && weapon.getItemLevel() < 15 && __1.QualifiedItemId == "(O)852" && !weapon.Name.Contains("Galaxy"))
+            {
+                if (Menu is not null && Menu.IsInnate && ReferenceEquals(Menu.Tool, weapon))
+                {
+                    __result = true;
+                    return false;
+                }
+                PossibleEnchantments.Clear();
+                foreach (BaseEnchantment enchantment in weapon.GetInnateEnchantments())
+                {
+                    PossibleEnchantments.Add(enchantment);
+                }
+                if (PossibleEnchantments.Count > 0)
+                {
+                    Menu = new(PossibleEnchantments, weapon, true);
                     __result = true;
                     return false;
                 }
@@ -120,6 +146,48 @@ internal static class ForgeMenuPatches
             ModEntry.ModMonitor.LogError("postfixing IsValidCraft", ex);
         }
         return true;
+    }
+
+    // Derived from attemptAddRandomInnateEnchantment
+    private static IEnumerable<BaseEnchantment> GetInnateEnchantments(this MeleeWeapon weapon)
+    {
+        int weaponLevel = weapon.getItemLevel();
+        if (weaponLevel <= 10)
+        {
+            yield return new DefenseEnchantment()
+            {
+                Level = Math.Clamp((Random.Shared.Next(weaponLevel + 1) / 2) + 1, 1, 2),
+            };
+        }
+
+        yield return new LightweightEnchantment()
+        {
+            Level = Random.Shared.Next(1, 6),
+        };
+
+        yield return new SlimeGathererEnchantment();
+
+        yield return new AttackEnchantment()
+        {
+            Level = Math.Clamp((Random.Shared.Next(weaponLevel + 1) / 2) + 1, 1, 5),
+        };
+
+        yield return new CritEnchantment
+        {
+            Level = Math.Clamp(Random.Shared.Next(weaponLevel) / 3, 1, 3),
+        };
+
+        yield return new WeaponSpeedEnchantment
+        {
+            Level = Math.Max(1, Math.Min(Math.Max(1, 4 - weapon.speed.Value), Random.Shared.Next(weaponLevel))),
+        };
+
+        yield return new SlimeSlayerEnchantment();
+
+        yield return new CritPowerEnchantment()
+        {
+            Level = Math.Clamp(Random.Shared.Next(weaponLevel) / 3, 1, 3),
+        };
     }
 
     /// <summary>