+++ StopRugRemoval/StopRugRemoval/HarmonyPatches/Niceties/KidsBeds.cs
@@ -1,0 +1,16 @@
+﻿using HarmonyLib;
+
+using StardewValley.Locations;
+using StardewValley.Objects;
+
+namespace StopRugRemoval.HarmonyPatches.Niceties;
+
+[HarmonyPatch(typeof(FarmHouse))]
+internal static class KidsBeds
+{
+    [HarmonyPatch(nameof(FarmHouse.GetChildBed))]
+    private static void Postfix(FarmHouse __instance, ref BedFurniture __result)
+    {
+        __result ??= __instance.GetBed(BedFurniture.BedType.Single);
+    }
+}